requirements.txt
Flask==3.1.0
pillow==11.2.1
scikit-learn
matplotlib==3.10.1
numpy==2.2.5
opencv-python==4.11.0.86
werkzeug==3.1.3
fpdf==1.7.2
